reorders a few sorting functions according to the intended use
