+++ packages/vizbuilder/src/helpers/sorting.js
@@ -1,6 +1,5 @@
 import sort from "fast-sort";
 import union from "lodash/union";
-import {unique} from "shorthash";
 
 /**
  * Checks if the dimension passed as argument is a time-type dimension.
@@ -16,63 +15,51 @@ export function isTimeDimension(dimension) {
 }
 
 /**
- * Prepares the array of cubes that will be used in the vizbuilder.
- * Specifically, filters the cubes that aren't for public use, and injects
- * information about the parent cube into the annotations of its measures
- * and levels.
- * @param {Cube[]} cubes An array of cubes. This array is modified in place.
+ * If `needle` is a valid value, returns the first element in the `haystack`
+ * that matches the annotation._key property.
+ * If there's no matches and `elseFirst` is true, returns the first element
+ * in the `haystack`.
+ * @param {string} needle The key to match
+ * @param {any[]} haystack The array where to search for the object.
+ * @param {boolean?} elseFirst A flag to return the first element in case of no matching result.
  */
-export function injectCubeInfoOnMeasure(cubes) {
-  let nCbs = cubes.length;
-  while (nCbs--) {
-    const cube = cubes[nCbs];
-    const cbAnnotations = cube.annotations;
+export function findByKey(needle, haystack, elseFirst = false) {
+  const findResult = needle ? haystack.find(item => item.annotations._key === needle) : undefined;
+  return elseFirst ? findResult || haystack[0] : findResult;
+}
 
-    if (cbAnnotations.hide_in_ui) {
-      cubes.splice(nCbs, 1);
-      continue;
-    }
+/**
+ * If `needle` is a valid value, returns the first element in the `haystack`
+ * that matches the name property.
+ * If there's no matches and `elseFirst` is true, returns the first element
+ * in the `haystack`.
+ * @param {string} needle The key to match
+ * @param {any[]} haystack The array where to search for the object.
+ * @param {boolean?} elseFirst A flag to return the first element in case of no matching result.
+ */
+export function findByName(needle, haystack, elseFirst = false) {
+  const findResult = needle ? haystack.find(item => item.name === needle) : undefined;
+  return elseFirst ? findResult || haystack[0] : findResult;
+}
 
-    const cbName = cube.caption || cube.name;
-    const cbTopic = cbAnnotations.topic || "Other";
-    const cbSubtopic = cbAnnotations.subtopic;
-    const selectorKey = `${cbTopic}_${cbSubtopic}_`;
-    const sourceName = cbAnnotations.source_name;
-
-    cbAnnotations._key = unique(cbName);
-
-    let nMsr = cube.measures.length;
-    while (nMsr--) {
-      const measure = cube.measures[nMsr];
-      const measureLabel = measure.caption || measure.name;
-      const msAnnotations = measure.annotations;
-
-      msAnnotations._key = unique(`${cbName} ${measure.name}`);
-      msAnnotations._cb_name = cbName;
-      msAnnotations._cb_topic = cbTopic;
-      msAnnotations._cb_subtopic = cbSubtopic;
-      msAnnotations._cb_sourceName = sourceName;
-      msAnnotations._selectorKey = selectorKey + measureLabel;
-    }
-
-    let nDim = cube.dimensions.length;
-    while (nDim--) {
-      const dimension = cube.dimensions[nDim];
-      const keyPrefix = `${cbName} ${dimension.name} `;
-
-      let nHie = dimension.hierarchies.length;
-      while (nHie--) {
-        const hierarchy = dimension.hierarchies[nHie];
-
-        let nLvl = hierarchy.levels.length;
-        while (nLvl--) {
-          const level = hierarchy.levels[nLvl];
-
-          level.annotations._key = unique(keyPrefix + level.name);
-        }
-      }
+/**
+ * Looks for an element of `defaults` in `haystack`, using the `matchingFunction`.
+ * If `elseFirst` is true and there's no match, returns the first element in `haystack`.
+ * @param {(needle, haystack) => any} matchingFunction The function to use to find the elements
+ * @param {any[]} haystack The array where to search for the object
+ * @param {string[]} defaults The array of default names to search for
+ * @param {boolean?} elseFirst A flag to return the first element in case of no matching result
+ */
+export function matchDefault(matchingFunction, haystack, defaults, elseFirst) {
+  let matchResult;
+  let n = defaults.length;
+  while (n--) {
+    const needle = defaults[n];
+    if (matchResult = matchingFunction(needle, haystack)) {
+      break;
     }
   }
+  return elseFirst ? matchResult || haystack[0] : matchResult;
 }
 
 /**
